Override styling with !important when necessary
There's already a PR out to update the TM UI for multiple interfaces,
which will supersede this change. I'm only doing this to make it easy
for testers to see the change in a cache server's availability, in case
that's not merged by the time the PR containing this commit is tested.