Fix bug converting interfaces to LID with multiple interfaces
When a set of interface info had only one of IPv4 or IPv6, and
non-service interfaces appeared after the service interface in the
iteration, the value stored at the interface name pointer would be
changed because the iteration variable is referential; therefore
although IP information was calculated correctly, the wrong
InterfaceName would be set.

This commit not only fixes that, but also actually speeds up the
conversion process in such cases.